SIP INVITE
SIP INVITE

Need to rearrange the code a bit the callee really should be the one to
give the SDP answer, it will be  needed for incoming SIP calls.

Also need a way to renew the REGISTER, scheduled action should do the
job just need to keep track of the sequence.